refactor: exported components and pages from their index.js file
