+++ src/App.js
@@ -5,7 +5,6 @@ import bcyrpt from "bcryptjs";
 import Mockman from "mockman-js";
 import { Route, Routes } from "react-router-dom";
 import Card from "./components/Card/Card";
-// import MockmanEs from "mockman-js";
 
 function App() {
   return <div className="App"></div>;
