+++ src/components/Navbar/Navbar.js
@@ -19,4 +19,4 @@ const Navbar = () => {
   );
 };
 
-export default Navbar;
+export { Navbar };
