+++ src/components/index.js
@@ -1,0 +1,3 @@
+export { Card } from "./Card/Card.js";
+export { FilterSideBar } from "./FilterSideBar/FilterSideBar.js";
+export { Navbar } from "./Navbar/Navbar.js";
