+++ src/pages/Login/Login.js
@@ -1,5 +1,5 @@
 import "./login.css";
-import Navbar from "../../components/Navbar/Navbar";
+import { Navbar } from "../../components";
 import { Link } from "react-router-dom";
 
 const Login = () => {
@@ -33,4 +33,4 @@ const Login = () => {
   );
 };
 
-export default Login;
+export { Login };
