+++ src/pages/index.js
@@ -1,0 +1,4 @@
+export { Login } from "./Login/Login.js";
+export { Signup } from "./Signup/Signup.js";
+export { ProductList } from "./productListing/ProductList.js";
+export { Wishlist } from "./wishlist/Wishlist.js";
