+++ src/pages/productListing/ProductList.js
@@ -1,12 +1,10 @@
 import axios from "axios";
 import { useEffect, useState } from "react";
-import Card from "../../components/Card/Card";
 
 import React from "react";
-import Navbar from "../../components/Navbar/Navbar";
-import FilterSideBar from "../../components/FilterSideBar/FilterSideBar";
 import useAxios from "../../CustomHooks/useAxios";
 import { useProducts } from "../../Context/product-context";
+import { Card, Navbar, FilterSideBar } from "../../components";
 
 const ProductList = () => {
   useAxios("/api/products", "get");
@@ -45,4 +43,4 @@ const ProductList = () => {
   );
 };
 
-export default ProductList;
+export { ProductList };
